fix: La réponse aux questions de personnalisation entraine un re-calcul des scores.
