+++ api_tests/tests/evaluation/actionStatut.test.ts
@@ -6,17 +6,11 @@ import {
 import { delay } from "https://deno.land/std@0.163.0/async/delay.ts";
 import { supabase } from "../../lib/supabase.ts";
 import { signIn, signOut } from "../../lib/auth.ts";
-import { ActionScore, ClientScores } from "../../lib/types/clientScores.ts";
+import { ClientScores } from "../../lib/types/clientScores.ts";
 import { ActionStatut } from "../../lib/types/actionStatut.ts";
 import { Avancement } from "../../lib/types/avancement.ts";
 import { testReset } from "../../lib/rpcs/testReset.ts";
-import { saveReponse } from "../../lib/rpcs/saveReponse.ts";
-
-function scoreById(clientScores: ClientScores, actionId: string): ActionScore {
-  return clientScores.scores.filter((s: ActionScore) =>
-    s.action_id === actionId
-  )[0]!;
-}
+import { scoreById } from "./scoreById.ts";
 
 Deno.test("Calcul des scores après la modification de statuts", async () => {
   await testReset();
@@ -70,53 +64,6 @@ Deno.test("Calcul des scores après la modification de statuts", async () => {
   assertEquals(actionScores2.point_programme, 0);
   assertNotEquals(actionScores2.point_fait, 0);
   assertEquals(actionScores2.point_fait, actionScores2.point_potentiel);
-
-  // on se déconnecte pour libérer les ressources
-  await signOut();
-});
-
-Deno.test("Calcul de 'eci_2.4.2' après la réponse à la question 'dechets_1'", async (t) => {
-  await testReset();
-  await signIn("yolododo");
-
-  const reponse = {
-    collectivite_id: 2,
-    question_id: "dechets_1",
-    reponse: false,
-  };
-  await saveReponse(reponse);
-
-  // on attend le calcul des scores
-  await delay(200);
-  let clientScores = await supabase.from<ClientScores>("client_scores")
-    .select()
-    .eq("collectivite_id", 2)
-    .eq("referentiel", "eci");
-
-  const eci242_1 = scoreById(clientScores.data![0], "eci_2.4.2");
-  assertEquals(
-    eci242_1.desactive,
-    true,
-    "Sans la compétence 'dechet_1' l'action devrait être désactivée",
-  );
-
-  // on change notre réponse
-  reponse.reponse = true;
-  await saveReponse(reponse);
-
-  // on attend de nouveau le calcul des scores
-  await delay(200);
-  clientScores = await supabase.from<ClientScores>("client_scores")
-    .select()
-    .eq("collectivite_id", 2)
-    .eq("referentiel", "eci");
-
-  const eci242_2 = scoreById(clientScores.data![0], "eci_2.4.2");
-  assertEquals(
-    eci242_2.desactive,
-    false,
-    "Avec la compétence 'dechet_1' l'action ne devrait pas être désactivée",
-  );
 
   // on se déconnecte pour libérer les ressources
   await signOut();
