+++ api_tests/tests/evaluation/scoreById.ts
@@ -1,0 +1,10 @@
+import { ActionScore, ClientScores } from "../../lib/types/clientScores.ts";
+
+export function scoreById(
+  clientScores: ClientScores,
+  actionId: string,
+): ActionScore {
+  return clientScores.scores.filter((s: ActionScore) =>
+    s.action_id === actionId
+  )[0]!;
+}
